Fix: sendMessageToFlutter를 util함수로 분리
