+++ components/liquor/category/CategoryHeader.tsx
@@ -4,25 +4,6 @@ import HeadBackIcon from "assets/icons/ico-head-back.svg";
 interface CategoryHeaderProps {
   tagValue: string;
 }
-// 앱으로 돌아가기
-const sendMessageToFlutter = () => {
-  console.log("Attempting to send message to Flutter...");
-
-  try {
-    // FlutterBridge 채널을 통해 메시지 전송
-    if (window.FlutterBridge) {
-      console.log("Flutter bridge detected, sending message...");
-      window.FlutterBridge.postMessage("goBack");
-      console.log("Message sent to Flutter successfully");
-    } else {
-      console.warn(
-        "No Flutter bridge detected. Are you running in a Flutter WebView?",
-      );
-    }
-  } catch (error) {
-    console.error("Error sending message to Flutter:", error);
-  }
-};
 
 // 카테고리 헤더 컴포넌트
 function CategoryHeader({ tagValue }: CategoryHeaderProps) {
